requirements.txt

authentication     == 1.1.0
fastapi             ==0.70.0
fastapi-cache       ==0.1.0
fastapi-redis-cache ==0.2.5
hiredis             ==2.0.0
jose                ==1.0.0
multipart           ==0.2.4
passlib             ==1.7.4
pip                 ==21.2.3
pydantic            ==1.8.2
pymongo             ==3.12.1
python-dateutil     ==2.8.2
python-dotenv       ==0.19.1
python-jose         ==3.3.0
python-multipart    ==0.0.5
redis               ==3.5.3
rsa                 ==4.7.2
SQLAlchemy          ==1.4.26
starlette           ==0.16.0
typing-extensions   ==3.10.0.2
uvicorn             ==0.15.0
vine               == 5.0.0
wcwidth            == 0.2.5
Werkzeug           == 2.0.2
WTForms            == 2.3.3
setuptools         == 57.4.0
six                == 1.16.0
sniffio            == 1.2.0
aioredis           == 1.3.1
amqp               == 5.0.6
anyio              == 3.3.4
asgiref            == 3.4.1
async-timeout      == 4.0.0